+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- link in playground quarkus hope page
+
 ## [3.1.6] - 2023-10-15
 
 ### Added